backend/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==3.2.3
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
drf-extra-fields==3.7.0
filetype==1.2.0
flake8==7.1.1
flake8-isort==6.1.1
idna==3.10
isort==5.13.2
mccabe==0.7.0
oauthlib==3.2.2
pep8-naming==0.14.1
Pillow==9.0.0
psycopg2-binary==2.9.3
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
